requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf_yasg==1.21.4
setuptools>=58.0.0
